# packages/pg-task-focus/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: pg-task-focus API
  version: 1.0.0
  summary: The loopback HTTP API of the pg-task-focus daemon.
  description: |
    One contract for every client: the command-line client, the web UI, the
    menu-bar plugin and the connector backend. It is spec-first: this document
    is the authoritative definition of every path, parameter, body and status,
    and a contract test fails when a registered route is missing from it or a
    path here has no route, and validates every response the server sends in
    the tests against it (except `/stream`, whose event format OpenAPI 3.1
    cannot describe and which is described in prose below).

    Errors are RFC 9457 `application/problem+json` with an extension member
    `reason`, a code from the closed set below, and a `trace_id`. Every
    condition has its own code; there is no catch-all such as an "invalid
    timeline". Statuses keep their HTTP meaning: 400 is a malformed or
    incomplete request (a cycle verb that does not say which of several
    cycles it means included), 404 a thing that does not exist, 409 a conflict
    with the current state or with other events, 422 a well-formed request
    whose instant or content is unacceptable for its own target, and 503 an
    unavailable store.

    The daemon listens on 127.0.0.1 only. Every route, `/healthz`, `/metrics`
    and `/stream` included, requires a `Host` header from the allowlist
...

...
    for inequality and refetches `/state` on any event.

...
servers:
...
paths:
  /:
    get:
      tags: [ops]
      operationId: getIndex
      summary: The page of the web UI (static HTML; it loads its modules and stylesheet from /assets/)
      responses:
        "200":
          {
            description: An HTML page,
            content: { text/html: { schema: { type: string } } },
          }
        default: { $ref: "#/components/responses/Problem" }
  /assets/{name}:
    get:
      tags: [ops]
      operationId: getAsset
      summary: One asset of the web UI, a script module or the stylesheet, embedded in the binary
      description: |
        The names are a fixed list built into the binary; any other name is a
        `not_found` problem. An asset is sent with its own content type, a strong
        `ETag` and `Cache-Control: no-cache`, so a new binary's page is seen on
        the next load and a cached one costs a conditional request (`304`). The
        page and its assets are sent under a content security policy that allows
        the page's own origin only and no inline script or style.
      parameters:
        - {
            name: name,
            in: path,
            required: true,
            schema: { type: string, minLength: 1 },
          }
      responses:
        "200":
          description: The asset.
          content:
            text/javascript: { schema: { type: string } }
            text/css: { schema: { type: string } }
        "304":
          {
            description: The asset is unchanged since the `If-None-Match` validator.,
          }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/state:
    get:
      tags: [state]
      operationId: getState
      summary: The state a client shows now
      responses:
        "200":
          {
            description: The state,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/State" } },
              },
          }
        "503": { $ref: "#/components/responses/NotReady" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/stream:
    get:
      tags: [state]
      operationId: getStream
      summary: Server-sent events for a change of the state version or the store
      responses:
        "200":
          description: An event stream (see the description of this document).
          content:
            text/event-stream:
              schema: { type: string }
        "503": { $ref: "#/components/responses/NotReady" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/periods/change:
    post:
      tags: [periods]
      operationId: changePeriods
      summary: Begin one or more periods as one atomic batch, or preview doing so
      requestBody: { $ref: "#/components/requestBodies/ChangePeriods" }
      responses:
        "200":
          description: The change, or its dry-run preview.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      oneOf:
                        [
                          { $ref: "#/components/schemas/Result" },
                          { $ref: "#/components/schemas/DryRunResult" },
                        ],
                    },
                },
            }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/profile/change:
    post:
      tags: [periods]
      operationId: changeProfile
      summary: Make another profile active, or preview doing so
      requestBody: { $ref: "#/components/requestBodies/ChangeProfile" }
      responses:
        "200":
          description: The change, or its dry-run preview.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      oneOf:
                        [
                          { $ref: "#/components/schemas/Result" },
                          { $ref: "#/components/schemas/DryRunResult" },
                        ],
                    },
                },
            }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/tasks/{id}/complete:
    post:
      tags: [tasks]
      operationId: completeTask
      summary: Complete a task
      parameters: [{ $ref: "#/components/parameters/TaskID" }]
      requestBody: { $ref: "#/components/requestBodies/CompleteTask" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/tasks/{id}/skip:
    post:
      tags: [tasks]
      operationId: skipTask
      summary: Skip a task, for a reason
      parameters: [{ $ref: "#/components/parameters/TaskID" }]
      requestBody: { $ref: "#/components/requestBodies/SkipTask" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/start:
    post:
      tags: [cycles]
      operationId: startCycle
      summary: Start a cycle; a running cycle is interrupted, not refused
      requestBody: { $ref: "#/components/requestBodies/StartCycle" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/pause:
    post:
      tags: [cycles]
      operationId: pauseCycle
      summary: Pause a cycle; pausing a paused cycle is a no-op
      requestBody: { $ref: "#/components/requestBodies/CycleVerb" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/resume:
    post:
      tags: [cycles]
      operationId: resumeCycle
      summary: Resume a paused cycle; resuming a running cycle is a no-op
      requestBody: { $ref: "#/components/requestBodies/CycleVerb" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/boost:
    post:
      tags: [cycles]
      operationId: boostCycle
      summary: Add minutes to a running or paused cycle
      requestBody: { $ref: "#/components/requestBodies/BoostCycle" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/stop:
    post:
      tags: [cycles]
      operationId: stopCycle
      summary: Stop a cycle; stopping a stopped cycle that is named is a no-op
      requestBody: { $ref: "#/components/requestBodies/CycleVerb" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/switch:
    post:
      tags: [cycles]
      operationId: switchCycle
      summary: Make a paused cycle the focus and pause the running one, as one batch
      requestBody: { $ref: "#/components/requestBodies/SwitchCycle" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/annotate:
    post:
      tags: [cycles]
      operationId: annotateCycle
      summary: Replace a cycle's note and key/value pairs, in any state
      requestBody: { $ref: "#/components/requestBodies/AnnotateCycle" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/cycles/break:
    post:
      tags: [cycles]
      operationId: breakCycle
      summary: Back-fill a break a cycle forgot, as one batch
      requestBody: { $ref: "#/components/requestBodies/BreakCycle" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/events:
    get:
      tags: [editor]
      operationId: listEvents
      summary: The logged events, as corrected or as first logged
      parameters:
        - {
            name: from,
            in: query,
            schema: { $ref: "#/components/schemas/Instant" },
            description: Inclusive lower bound on the effective instant in the chosen view.,
          }
        - {
            name: to,
            in: query,
            schema: { $ref: "#/components/schemas/Instant" },
            description: Exclusive upper bound.,
          }
        - {
            name: type,
            in: query,
            style: form,
            explode: true,
            schema:
              {
                type: array,
                items: { $ref: "#/components/schemas/EventType" },
              },
            description: Event types to keep; repeatable or comma separated.,
          }
        - {
            name: view,
            in: query,
            schema: { enum: [corrected, original], default: corrected },
          }
      responses:
        "200":
          {
            description: The events,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Events" } },
              },
          }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/events/{id}/correct:
    post:
      tags: [editor]
      operationId: correctEvent
      summary: Replace fields of a stored event; the original stays in the log
      parameters: [{ $ref: "#/components/parameters/EventID" }]
      requestBody: { $ref: "#/components/requestBodies/Correct" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/events/{id}/retract:
    post:
      tags: [editor]
      operationId: retractEvent
      summary: Take back one stored event (undo)
      parameters: [{ $ref: "#/components/parameters/EventID" }]
      requestBody: { $ref: "#/components/requestBodies/Retract" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/batches/{id}/retract:
    post:
      tags: [editor]
      operationId: retractBatch
      summary: Take back a whole batch (undo)
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/ULID" },
          },
        ]
      requestBody: { $ref: "#/components/requestBodies/Retract" }
      responses:
        "200": { $ref: "#/components/responses/Mutation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/config:
    get:
      tags: [state]
      operationId: getConfig
      summary: The resolved, validated configuration
      responses:
        "200":
          {
            description: The configuration,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ConfigDoc" } },
              },
          }
        "503": { $ref: "#/components/responses/NotReady" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/calendar:
    get:
      tags: [connector]
      operationId: getCalendar
      summary: The running segments of cycles that overlap a window
      parameters:
        - {
            name: from,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/Instant" },
          }
        - {
            name: to,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/Instant" },
          }
        - {
            name: calendar,
            in: query,
            schema: { type: string },
            description: Empty or the fixed calendar id or name; any other name returns no events.,
          }
      responses:
        "200":
          {
            description: The segments,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Calendar" } },
              },
          }
        default: { $ref: "#/components/responses/Problem" }
  /api/v1/attention:
    get:
      tags: [connector]
      operationId: getAttention
      summary: The attention feed, stateless
      responses:
        "200":
          {
            description: The items,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Attention" } },
              },
          }
        "503": { $ref: "#/components/responses/NotReady" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  parameters:
    TaskID:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1 }
      description: A task id, `<day|week|sprint>:<period start>:<definition>`.
...
  responses:
...
    Problem:
      description: A refusal, RFC 9457. The `reason` is from the closed set.
      content:
        {
          application/problem+json:
            { schema: { $ref: "#/components/schemas/Problem" } },
        }
    NotReady:
      description: Replay has not finished or the first store check has not passed (`not_ready`).
      content:
        {
          application/problem+json:
            { schema: { $ref: "#/components/schemas/Problem" } },
        }
  requestBodies:
    ChangePeriods:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ChangePeriodsRequest" } },
        }
    ChangeProfile:
      required: true
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ChangeProfileRequest" } },
        }
...
  schemas:
...
    State:
      type: object
      required:
        [
          read_at,
          initialized,
          version,
          store,
          periods,
          tasks,
          next,
          focus,
          dimmed,
          interrupt_stack,
          resume_offer,
        ]
      additionalProperties: false
      properties:
        read_at: { $ref: "#/components/schemas/Instant" }
        initialized: { type: boolean }
        profile: { type: string }
        version: { $ref: "#/components/schemas/Version" }
        store: { $ref: "#/components/schemas/Store" }
        periods: { type: array, items: { $ref: "#/components/schemas/Period" } }
        tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
        next:
          { oneOf: [{ $ref: "#/components/schemas/Task" }, { type: "null" }] }
        focus:
          { oneOf: [{ $ref: "#/components/schemas/Cycle" }, { type: "null" }] }
        dimmed:
          {
            type: array,
            items: { $ref: "#/components/schemas/DimmedCycle" },
            description: Every paused cycle that is not stopped,
            most recently paused first.,
          }
        interrupt_stack:
          { type: array, items: { $ref: "#/components/schemas/Cycle" } }
        resume_offer:
          {
            oneOf:
              [{ $ref: "#/components/schemas/ResumeOffer" }, { type: "null" }],
          }
...